requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.25.0
tldextract>=5.1.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
openpyxl>=3.1.0
